Support (prefer) filepaths for cohort/label queries
Maintaining a large project with multiple experiment files can become
burdensome when the same cohort or label queries are used in multiple
experiments. This is made easier by allowing users to include filenames
in place of code in the config. Triage loads the query into the config
before hashing components or the experiment to ensure that the query
itself is versioned, rather than the filepath.