Support (prefer) filepaths for cohort/label queries (#883)
* Support (prefer) filepaths for cohort/label queries

Maintaining a large project with multiple experiment files can become
burdensome when the same cohort or label queries are used in multiple
experiments. This is made easier by allowing users to include filenames
in place of code in the config. Triage loads the query into the config
before hashing components or the experiment to ensure that the query
itself is versioned, rather than the filepath.

* Default to query from file in tests

* Fix failing test

* Fix experiment tests

* Fix architect integration tests

* Pass validation with filepaths

* Add unit tests for loading queries from files

* Fix postmodeling tests

* Move query loader to triage.util.conf

* Condense with statements

* Copy instead of side effect

* remove swap files :facepalm:

* more explicit