use >=, not <=
In general, we try to use `>=` and `<` rather than `<=` or `>` for time inequalities in our examples to be consistent with the fact that triage internally uses closed-right-ends and open-left-ends for time intervals. This makes translating examples into triage behavior easier. This commit fixes an example where we used `<=` for label creation.